Organize storybook categories
