+++ stories/explorer/index.stories.js
@@ -1,6 +1,6 @@
 import 'babel-polyfill'
 import { storiesOf } from '@storybook/react'
 
-storiesOf('Index Explorer', module)
+storiesOf('Components (Unthemed)|Index Explorer', module)
   .addWithJSX('Advanced Search', require('./advanced').default)
   .addWithJSX('Basic Search', require('./basic').default)
