+++ stories/greyVest/box.stories.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { Box, Fonts, GVStyle } from './../../src/themes/greyVest'
 
-storiesOf('Non Search Components|Grey Vest', module).addWithJSX('Box', () => (
+storiesOf('Components (Grey Vest)|Box', module).addWithJSX('Default', () => (
   <div>
     <Fonts />
     <GVStyle />
