+++ stories/greyVest/linkButton.stories.js
@@ -3,8 +3,8 @@ import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
 import { LinkButton, Fonts, GVStyle } from './../../src/themes/greyVest'
 
-storiesOf('Non Search Components|Grey Vest', module).addWithJSX(
-  'LinkButton',
+storiesOf('Components (Grey Vest)|LinkButton', module).addWithJSX(
+  'Default',
   () => (
     <div>
       <Fonts />
