+++ stories/queryBuilder/internals/addPreview.js
@@ -4,7 +4,7 @@ import { action } from '@storybook/addon-actions'
 import AddPreview from '../../../src/queryBuilder/preview/AddPreview'
 
 export default () =>
-  storiesOf('QueryBuilder/Internals/AddPreview', module)
+  storiesOf('Search Components (Unthemed)|QueryBuilder/Internals/AddPreview', module)
     .addWithJSX('and', () => <AddPreview onClick={action('join')} join="and" />)
     .addWithJSX('or', () => <AddPreview onClick={action('join')} join="or" />)
     .addWithJSX('not', () => <AddPreview onClick={action('join')} join="not" />)
