+++ stories/queryBuilder/internals/group.js
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react'
 import Group from '../../../src/queryBuilder/Group'
 
 export default (parent, root, DnDDecorator) =>
-  storiesOf('QueryBuilder/Internals/Group', module)
+  storiesOf('Search Components (Unthemed)|QueryBuilder/Internals/Group', module)
     .addDecorator(DnDDecorator)
     .addWithJSX('One Filter', () => (
       <Group
