+++ stories/queryBuilder/internals/indentable.js
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react'
 import Indentable from '../../../src/queryBuilder/preview/Indentable'
 
 export default () =>
-  storiesOf('QueryBuilder/Internals/Indentable', module)
+  storiesOf('Search Components (Unthemed)|QueryBuilder/Internals/Indentable', module)
     .addWithJSX('and', () => (
       <Indentable indent={() => true} tree={{ join: 'and' }}>
         <div style={{ height: '100px' }}>Contents</div>
