+++ stories/queryBuilder/internals/operatorMenu.js
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react'
 import OperatorMenu from '../../../src/queryBuilder/OperatorMenu'
 
 export default (parent, root) =>
-  storiesOf('QueryBuilder/Internals/OperatorMenu', module).addWithJSX(
+  storiesOf('Search Components (Unthemed)|QueryBuilder/Internals/OperatorMenu', module).addWithJSX(
     'OperatorMenu',
     () => <OperatorMenu {...{ tree: { join: 'and' }, parent, root }} />
   )
